backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.1.0
pydantic>=2.0.0
requests>=2.31.0
feedparser>=6.0.0
